requirements.txt
django==3.2.5
djangorestframework==3.11.1
djangorestframework-simplejwt==4.6.0
psycopg2-binary==2.8.6
python-decouple==3.3